Soft Delete of APIRevisions working

+++ src/dotnet/APIView/ClientSPA/src/app/_components/revisions-list/revisions-list.component.ts
@@ -1,6 +1,6 @@
 import { AfterViewInit, Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
 import { MenuItem, SortEvent } from 'primeng/api';
-import { TableFilterEvent, TableLazyLoadEvent } from 'primeng/table';
+import { Table, TableFilterEvent, TableLazyLoadEvent } from 'primeng/table';
 import { Pagination } from 'src/app/_models/pagination';
 import { Review } from 'src/app/_models/review';
 import { Revision } from 'src/app/_models/revision';
@@ -20,15 +20,17 @@ export class RevisionsListComponent implements OnInit, OnChanges {
   totalNumberOfRevisions = 0;
   pagination: Pagination | undefined;
   insertIndex : number = 0;
-  rowHeight: number = 88;
+  rowHeight: number = 48;
   noOfRows: number = Math.floor((window.innerHeight * 0.75) / this.rowHeight); // Dynamically Computing the number of rows to show at once
   pageSize = 20; // No of items to load from server at a time
+  sortField : string = "lastUpdatedOn";
+  sortOrder : number = 1;
+  filters: any = null;
 
   // Filters
   details: any[] = [];
   selectedDetails: any[] = [];
   showDeletedAPIRevisions : boolean = false;
-
   sidebarVisible : boolean = false;
 
   // Context Menu
@@ -37,6 +39,9 @@ export class RevisionsListComponent implements OnInit, OnChanges {
   selectedRevisions!: Revision[];
   showSelectionActions : boolean = false;
   showDiffButton : boolean = false;
+
+  // Messages
+  apiRevisionsListDetail: string = "APIRevisions in"
 
   badgeClass : Map<string, string> = new Map<string, string>();
 
@@ -73,7 +78,7 @@ export class RevisionsListComponent implements OnInit, OnChanges {
       details = (filters.details.value != null) ? filters.details.value.map((item: any) => item.data): details;
     }
 
-    this.revisionsService.getRevisions(noOfItemsRead, pageSize, reviewId, label, author, details, sortField, sortOrder).subscribe({
+    this.revisionsService.getAPIRevisions(noOfItemsRead, pageSize, reviewId, label, author, details, sortField, sortOrder, this.showDeletedAPIRevisions).subscribe({
       next: response => {
         if (response.result && response.pagination) {
           if (resetReviews)
@@ -81,6 +86,8 @@ export class RevisionsListComponent implements OnInit, OnChanges {
             const arraySize = Math.ceil(response.pagination!.totalCount + Math.min(20, (0.05 * response.pagination!.totalCount))) // Add 5% extra rows to avoid flikering
             this.revisions = Array.from({ length: arraySize });
             this.insertIndex = 0;
+            this.showSelectionActions = false;
+            this.showDiffButton = false;
           }
 
           if (response.result.length > 0)
@@ -96,10 +103,19 @@ export class RevisionsListComponent implements OnInit, OnChanges {
   }
 
   createContextMenuItems() {
-    this.contextMenuItems = [
-      { label: 'View', icon: 'pi pi-fw pi-search', command: () => this.viewRevision(this.selectedRevision) },
-      { label: 'Delete', icon: 'pi pi-fw pi-times', command: () => this.deleteRevision(this.selectedRevision) }
-    ];
+    if (this.showDeletedAPIRevisions)
+    {
+      this.contextMenuItems = [
+        { label: 'Restore', icon: 'pi pi-folder-open', command: () => this.viewRevision(this.selectedRevision) }
+      ];
+    }
+    else 
+    {
+      this.contextMenuItems = [
+        { label: 'View', icon: 'pi pi-folder-open', command: () => this.viewRevision(this.selectedRevision) },
+        { label: 'Delete', icon: 'pi pi-fw pi-times', command: () => this.deleteRevision(this.selectedRevision) }
+      ];
+    }
   }
 
   createFilters() {     
@@ -141,11 +157,62 @@ export class RevisionsListComponent implements OnInit, OnChanges {
   }
   
   viewRevision(revision: Revision) {
+    if (!this.showDeletedAPIRevisions)
+    {
       this.revisionsService.openAPIRevisionPage(this.review!.id, revision.id);
+    }
+  }
+
+  deleteRevisions(revisions: Revision []) {
+    this.revisionsService.deleteAPIRevisions(this.review!.id, revisions.map(r => r.id)).subscribe({
+      next: response => {
+        if (response) {
+          this.loadRevisions(0, this.pageSize * 2, true);
+        }
+      }
+    });
   }
 
   deleteRevision(revision: Revision) {
-      
+    this.revisionsService.deleteAPIRevisions(revision.reviewId, [revision.id]).subscribe({
+      next: response => {
+        if (response) {
+          this.loadRevisions(0, this.pageSize * 2, true);
+        }
+      }
+    });
+  }
+
+  /**
+  * Return true if table has filters applied.
+  */
+  tableHasFilters() : boolean {
+    return (this.sortField != "lastUpdatedOn" || this.sortOrder != 1 || (this.filters && (this.filters.label.value != null || this.filters.author.value != null || this.filters.details.value != null)));
+  }
+
+  /**
+  * Clear all filters in Table
+  */
+  clear(table: Table) {
+    table.clear();
+    this.loadRevisions(0, this.pageSize * 2, true);
+  }
+
+  /**
+  * Toggle Show deleted APIRevisions
+  */
+  toggleShowDeletedAPIRevisions() {
+    this.showDeletedAPIRevisions = !this.showDeletedAPIRevisions;
+    this.loadRevisions(0, this.pageSize * 2, true);
+    this.createContextMenuItems();
+    if (!this.showDeletedAPIRevisions)
+    {
+      this.apiRevisionsListDetail = "APIRevisions";
+    }
+    else
+    {
+      this.apiRevisionsListDetail = "Deleted APIRevisions in";
+    }
   }
 
   /**
@@ -155,13 +222,14 @@ export class RevisionsListComponent implements OnInit, OnChanges {
   onLazyLoad(event: TableLazyLoadEvent) {
       console.log("On Lazy Event Emitted %o", event);
       const last = Math.min(event.last!, this.totalNumberOfRevisions);
+      this.sortField = event.sortField as string ?? "lastUpdatedOn";
+      this.sortOrder = event.sortOrder as number ?? 1;
+      this.filters = event.filters;
       if (last! > (this.insertIndex - this.pageSize))
       {
         if (this.pagination && this.pagination?.noOfItemsRead! < this.pagination?.totalCount!)
         {
-          const sortField : string = event.sortField as string ?? "lastUpdated";
-          const sortOrder : number = event.sortOrder as number ?? 1;
-          this.loadRevisions(this.pagination!.noOfItemsRead, this.pageSize, false, event.filters, sortField, sortOrder);
+          this.loadRevisions(this.pagination!.noOfItemsRead, this.pageSize, false, event.filters, this.sortField, this.sortOrder);
         }
       }
       event.forceUpdate!();
